alarm clock improvements feature: added days of week to alarm planning and a sound beep while an alarm occurs
refactoring: refactored way how are objects placed to tiles - offseting was replaced by placing to containers with a layout and padding. The padding is used as an extended clickable area. It lead to decreasing number of wrong clicking to an clickable object. It lead to unification of dialog layouts.

refactorig: rtc corresponding alarm functionality was fully moved to rtcctl
including its storing to a persistent file. Alarm occurrence corresponding functionality was moved to alarm_clock (alarm_clock_data was removed)

improvement: roller object in alarm_clock_main was adjusted to be
operable more smoothly (decreased height of line + addend 1 visible line)