Reorder logs of blectl
It is easier to read if log is emitted as soon as it is concerned
and at least before blectl_send_event_cd.
As blectl_send_event_cd effectively run the registered clients,
if the log is emitted after this call we read the log of the event
after the log of its consequences.